+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `depth` parameter added to `save_image` method
+- `depth` parameter added to `save_fig_file` function
 ## [0.2] - 2021-10-14
 ### Added
 - `dependabot.yml`